made initial split for power_suppply mode
